fix(evaluation): deduplicate annotation labels component to pass Sonar quality gate
